requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
psycopg[binary]==3.2.1
passlib[bcrypt]==1.7.4
